Move color area gradients inside aria hook like other color components
