+++ packages/@react-aria/color/src/index.ts
@@ -14,4 +14,3 @@ export * from './useColorArea';
 export * from './useColorSlider';
 export * from './useColorWheel';
 export * from './useColorField';
-export * from './gradients';
